chore(dsystemtime): Changing the behavior of some functions
改变部分函数行为

Log: 改变部分函数行为和重命名部分函数

+++ dtksystemtime/tests/fakedbus/timedate1service.h
@@ -43,7 +43,7 @@ public slots:
 public:
     bool setLocalRTC_sig = false;
     bool setLocalTime_sig = false;
-    int setTime_sig = 0;
+    qint64 setTime_sig = 0;
     QString setTimeZone_sig;
 
 private:
